[1.9.1.3] Changed - GA and PSO: Early convergence detection now based on the mean coefficient of variation of the kinetic parameters, relative to that of the initial population. - "hybrid_TSA" option: add two extra steps to the original DRG (or SA) algorithm to combine it with concentration ranking reduction (CRR)
[1.9.1.2]
Added:
- "1 by 1 addition": Reduction iteration now can go species-by-species when the error threshold is exceeded (default option)

Fixed
- sensitivity analysis (SA) on the reduced mechanism when mech_prev_red is used
- Fixed convergence issues in calculations during optimization via GA/PSO
- Corrected consideration of proxy species for flame speed/ignition delay time/stretch for the identification of reactions to optimize via DRG analysis

[1.9.1.1]
Changed
- Improved time discretization for reactor simulations

[1.9.1.0]
Added
- select_wd(<path>) function to manage working directories without using the GUI.

Changed
- Adaptation of the model writing tool (YAML → Chemkin).
- Improved management of lower and upper bounds for optimization based on uncertainties (offset for very small values).
- Improved computation of optimized models (removed reading of the reference mechanism by Cantera previously used as a species index dictionary).

+++ brookesia/Opt_tools.py
@@ -108,7 +108,7 @@ def check_k(_react,_r):
             At = np.float32(_react.param[_r]['A'])
             T3t = np.float32(_react.param[_r]['T3'])
             T1t = np.float32(_react.param[_r]['T1'])
-            T2t = np.float32(_react.param[_r]['T2'])
+            T2t = np.float64(_react.param[_r]['T2'])
         except:
             Troe_func = False
         # calculation of k
